--- library/src/androidTest/java/org/xmtp/android/library/IdentityTest.kt
@@ -1,64 +1,0 @@
-package org.xmtp.android.library
-
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.platform.app.InstrumentationRegistry
-import kotlinx.coroutines.runBlocking
-import org.junit.Before
-import org.junit.Test
-import org.junit.runner.RunWith
-import org.xmtp.android.library.messages.PrivateKey
-import org.xmtp.android.library.messages.PrivateKeyBuilder
-import org.xmtp.android.library.messages.walletAddress
-
-@RunWith(AndroidJUnit4::class)
-class IdentityTest {
-    private lateinit var alixWallet: PrivateKeyBuilder
-    private lateinit var boWallet: PrivateKeyBuilder
-    private lateinit var alix: PrivateKey
-    private lateinit var alixClient: Client
-    private lateinit var bo: PrivateKey
-    private lateinit var boClient: Client
-    private lateinit var caroWallet: PrivateKeyBuilder
-    private lateinit var caro: PrivateKey
-    private lateinit var caroClient: Client
-    private lateinit var fixtures: Fixtures
-
-    @Before
-    fun setUp() {
-        val context = InstrumentationRegistry.getInstrumentation().targetContext
-        fixtures =
-            fixtures(
-                clientOptions = ClientOptions(
-                    ClientOptions.Api(XMTPEnvironment.LOCAL, false),
-                    enableAlphaMls = true,
-                    appContext = context
-                )
-            )
-        alixWallet = fixtures.aliceAccount
-        alix = fixtures.alice
-        boWallet = fixtures.bobAccount
-        bo = fixtures.bob
-        caroWallet = fixtures.caroAccount
-        caro = fixtures.caro
-
-        alixClient = fixtures.aliceClient
-        boClient = fixtures.bobClient
-        caroClient = fixtures.caroClient
-    }
-
-    @Test
-    fun testCanAllowAndDenyInboxId() {
-        assert(!boClient.contacts.isInboxIdAllowed(alixClient.inboxId))
-        assert(!boClient.contacts.isInboxIdDenied(alixClient.inboxId))
-
-        runBlocking { boClient.contacts.allowInboxId(listOf(alixClient.inboxId)) }
-
-        assert(boClient.contacts.isInboxIdAllowed(alixClient.inboxId))
-        assert(!boClient.contacts.isInboxIdDenied(alixClient.inboxId))
-
-        runBlocking { boClient.contacts.denyInboxIds(listOf(alixClient.inboxId)) }
-
-        assert(!boClient.contacts.isInboxIdAllowed(alixClient.inboxId))
-        assert(boClient.contacts.isInboxIdDenied(alixClient.inboxId))
-    }
-}
